+++ X-RayPalette/Views/InfoChange/WrapperInfoChange.cs
@@ -1,0 +1,28 @@
+﻿using ImGuiNET;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace X_RayPalette.Views.InfoChange
+{
+    public class WrapperInfoChange : View
+    {
+        public override void Back()
+        {
+            OnBackEvent();
+        }
+
+        public override void Render(bool isAdmin)
+        {
+            ImGui.PushStyleColor(ImGuiCol.Button, new Vector4(0.0f, 0.749f, 1.0f, 0.20f));
+            if (ImGui.Button("< back"))
+            {
+                Back();
+            }
+            ImGui.PopStyleColor();
+        }
+    }
+}
